+++ CHANGELOG.md
@@ -3,18 +3,29 @@
 # Exam Plugin Changelog
 
 ## [Unreleased]
+
+## [0.2.0] - 2024-01-12
+
 ### Added
+
 - Add spec files templates #11
 
 ## [0.1.0]
+
 ### Added
+
 - Add Exam file type
 - Add Debug button
 
 ### Changed
+
 - Add Exam tool window straightway
 - Change markdown example name to header text
 
 ### Fixed
+
 - Invalid test name for markdown example name matching other header #8
 
+[Unreleased]: https://github.com/MetallFoX/ExamPlugin/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/MetallFoX/ExamPlugin/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/MetallFoX/ExamPlugin/commits/v0.1.0